requirements.txt
asgiref==3.4.1
Django==2.2.10
django-filter==21.1
django-mssql==1.8
django-mssql-backend==2.8.1
djangorestframework==3.13.1
importlib-metadata==4.10.1
Markdown==3.3.6
Pillow==9.0.0
pyodbc==4.0.32
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
zipp==3.7.0
